Finally did renaming of material2 to material
