--- Assets/LeapMotionModules/InteractionEngine/Scripts/Editor/InteractionMaterial2Editor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8c56828192eb7444b87de3b880560f60
-timeCreated: 1464224087
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
