+++ Assets/LeapMotionModules/InteractionEngine/Scripts/Editor/InteractionMaterialEditor.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 25640ad07d6be1f42bec4bfe28c60aae
-timeCreated: 1462498436
+guid: 8c56828192eb7444b87de3b880560f60
+timeCreated: 1464224087
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
