--- Assets/LeapMotionModules/InteractionEngine/Scripts/InteractionMaterial2.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 9e240ddf8ae2bd147b6c7d3cd1da1e2a
-timeCreated: 1464223744
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
